Added repository for entries, upgraded Summary Menu

+++ TrackIt/UI/Model/Entry.cs
@@ -4,7 +4,7 @@
     {
         public bool Type; //If its true its Expense, if its false its Income.
         public string Description;
-        public int Amount;
+        public decimal Amount;
         public DateTime Date;
         public string Category;
         public string? Period; //If (Period = null) it's a unique Expense/Income entry.
